refactor: migrate runtime to typescript
