+++ packages/types/cli/index.d.ts
@@ -1,0 +1,8 @@
+import { Configuration } from '../config'
+
+type Command = any // TBD
+
+export interface Hooks {
+  config?(config: Configuration): void
+  'run:before'?(params: { argv: string [], command: Command, rootDir: string }): void
+}
