+++ packages/typescript-runtime/package.json
@@ -6,13 +6,18 @@
   "license": "MIT",
   "files": [
     "bin",
-    "lib"
+    "dist"
   ],
   "bin": {
     "nuxt-ts": "bin/nuxt-ts.js",
     "nuxts": "bin/nuxt-ts.js"
   },
-  "main": "lib/index.js",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "yarn run clean && yarn compile",
+    "clean": "rm -rf dist",
+    "compile": "tsc"
+  },
   "dependencies": {
     "ts-node": "^8.4.1",
     "typescript": "~3.6"
