+++ packages/typescript-runtime/test/hooks.test.ts
@@ -1,16 +1,17 @@
 import path from 'path'
 import { register } from 'ts-node'
-import { hooks } from '..'
+import { Configuration as NuxtConfiguration } from '@nuxt/types'
+import { hooks } from '../src'
 
 jest.mock('ts-node')
 
 describe('run:before hook', () => {
   beforeEach(() => {
-    register.mockClear()
+    jest.clearAllMocks()
   })
 
   test('registers ts-node', () => {
-    hooks['run:before']({ argv: [], rootDir: 'path' })
+    hooks['run:before']!({ argv: [], command: null, rootDir: 'path' })
 
     expect(register).toHaveBeenCalledWith({
       project: path.resolve('path', 'tsconfig.json'),
@@ -21,7 +22,7 @@ describe('run:before hook', () => {
   })
 
   test('registers ts-node (custom tsconfig.json path)', () => {
-    hooks['run:before']({ argv: ['--tsconfig', 'custom/tsconfig.json'], rootDir: 'path' })
+    hooks['run:before']!({ argv: ['--tsconfig', 'custom/tsconfig.json'], command: null, rootDir: 'path' })
 
     expect(register).toHaveBeenCalledWith({
       project: path.resolve('custom/tsconfig.json'),
@@ -32,7 +33,7 @@ describe('run:before hook', () => {
   })
 
   test('registers ts-node (custom tsconfig.json dir path)', () => {
-    hooks['run:before']({ argv: ['--tsconfig', 'custom'], rootDir: 'path' })
+    hooks['run:before']!({ argv: ['--tsconfig', 'custom'], command: null, rootDir: 'path' })
 
     expect(register).toHaveBeenCalledWith({
       project: path.resolve('custom/tsconfig.json'),
@@ -47,15 +48,15 @@ describe('config hook', () => {
   test('adds ts extension (config.extensions is: defined)', () => {
     const config = { extensions: [] }
 
-    hooks.config(config)
+    hooks.config!(config)
 
     expect(config.extensions).toContain('ts')
   })
 
   test('adds ts extension (config.extensions is: undefined)', () => {
-    const config = {}
+    const config: NuxtConfiguration = {}
 
-    hooks.config(config)
+    hooks.config!(config)
 
     expect(config.extensions).toContain('ts')
   })
