+++ packages/typescript-runtime/tsconfig.json
@@ -1,9 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true,
-    "sourceMap": true
+    "module": "commonjs",
+    "outDir": "dist"
   },
   "include": [
     "src/**/*"
