+++ tsconfig.json
@@ -2,8 +2,6 @@
   "compilerOptions": {
     "target": "esnext",
     "lib": ["dom", "esnext"],
-    "declaration": true,
-    "sourceMap": true,
     "module": "esnext",
     "moduleResolution": "node",
     "skipLibCheck": true,
